Game: Moved completely to static .vox assets for WG
Turns out building a tree in goxel can be done in a couple
of minutes, and not just by me but by pretty much anybody. This
is also are super convenient way to allow the worldgenerator to
be extended by mods, since a mod could just add another tree to
the Vec.